+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 8.3.0
+
+- Add `validDropSquares` to `GameData`. For drop moves, `Chessboard` will now only call `onMove` 
+  if and only if `validDropSquares` contains the target square.
+  This makes the behavior consistent with normal moves.
+  Note that dropping a pawn on the back rank does not trigger `onMove`, even if `validDropSquares` would allow it.
+
 ## 8.2.1
 
 - Fix top player promotion selector when `canPromoteToKing` is true.